requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf
python-multipart
openai
tavily-python
openpyxl
pandas
langgraph
python-docx
